Removes optional AS <path pattern> from matchrules
Returns to mapping source artifacts to target artifacts stricly
by full filepaths instead of just by hashes, which breaks in case
of multiple different files with the same hash.

Matchrules used to have an optional second path pattern parameter,
to allow path and file renames outside of in-toto steps.
In combination with wildcards it became impossible to reliably
map source artifacts to target artifacts in order to match their
hashes.
https://github.com/in-toto/in-toto/issues/43#issuecomment-267472109